requirements.txt
python-telegram-bot[job-queue]>=20.0
pytz>=2023.3
anthropic>=0.40.0
